Improve search ID lookup

+++ trakt-api/src/commonMain/kotlin/app/moviebase/trakt/model/TraktRatingModel.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class TraktRating(
     val rating: Double,
-    val votes: Double
+    val votes: Int
 )
